pyproject.toml
[tool.pdm]

[project]
name = "Matrix-OrderBot"
version = "0.1.0"
description = ""
authors = [
    { name = "Zayus13", email = "sven_loechte@protonmail.com" },
]
dependencies = ["matrix-nio==0.19.0", "SQLAlchemy==1.4.37"]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[project.scripts]
orderbot = "orderbot.orderbot:main"